refactor: Button Props type 별도 파일로 분리
